#56 extract interface definitions into its own file
